org_service/requirements.txt
Django==5.2.1
djangorestframework==3.15.1
django-cors-headers==4.3.1
coverage==7.3.2
drf-yasg==1.21.7
python-dotenv==1.0.1
requests==2.31.0